Fix blog post route: layout with Outlet and /blog/ index
TanStack Router nested /blog/$slug under /blog without an Outlet, so post
pages never rendered. Split list UI into blog/index.tsx.
